5.13.10.6 for 08/17/12 - Updated opcodes/structs - Added Hero's Forge armor support changes
Version to 5.13.10.6


git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/trunk@769 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/filter.cpp
@@ -56,8 +56,8 @@ protected:
 /* FilterItem Class - allows easy creation / deletion of regex types */
 FilterItem::FilterItem(const QString& filterPattern, bool caseSensitive)
 {
-  uint8_t minLevel = 0;
-  uint8_t maxLevel = 0;
+  uint32_t minLevel = 0;
+  uint32_t maxLevel = 0;
   
   QString workString = filterPattern;
